refactor: rename from users. to user.
