+++ src/user/user.controller.ts
@@ -13,7 +13,7 @@ import {
   ClassSerializerInterceptor,
 } from '@nestjs/common';
 
-import { UsersService } from './users.service';
+import { UsersService } from './user.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
