+++ src/user/user.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { UsersService } from './users.service';
-import { UsersController } from './users.controller';
+import { UsersService } from './user.service';
+import { UsersController } from './user.controller';
 
 @Module({
   controllers: [UsersController],
